Fix case-(in)sensitivity directory name issues
Having both a "GUI" and a "gui" directory in the source tree causes
errors when trying to check out a working copy on a case-insensitive
filesystem. Move files to fix it.

There still appear to be inconsistencies in casing where project files
refer to both "GUI" and "gui" but at least the filesystem is consistent
now.